[Streamline] Clean repo and remove usedless file.
 - Remove nv plugin interface and dlls.
 - Change build and package scripts, only build/package mtss-fg releted file.

+++ source/plugins/sl.mtss_g/debugoverlay.cpp
@@ -115,7 +115,7 @@ void ImGuiDebugOverlay::Init(
 
         m_pImGuiCtx = m_pImGui->createContext(desc);
         m_pImGui->setCurrentContext(m_pImGuiCtx);
-        m_pImGui->setDisplaySize(sl::type::Float2(width, height));
+        m_pImGui->setDisplaySize(sl::type::Float2(static_cast<float>(width), static_cast<float>(height)));
 
         bool success = CreateInternalPixelShader();
         assert(success == true);
